feat(ui): update layout
